--- Arcano/composer.json
@@ -1,19 +1,0 @@
-{
-    "name": "arcanos/enigmas",
-    "description": "Projeto Arcanos",
-    "type": "project",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Everton Cardoso",
-            "email": "everton_cardoso@live.com"
-        }
-    ],
-    "require": {},
-    "autoload": {
-        "psr-4": {
-            "Arcanos\\Enigmas\\" : "src/"
-        },
-        "classmap": ["./config/Routes.php"]
-    }
-}
