--- Arcano/public/css/dropdown.css
@@ -1,209 +1,0 @@
-@import 'https://fonts.googleapis.com/css?family=Raleway';
-
-.profile {
-  width: 226px;
-  height: 48px;
-  transform: translate(0%, 0%);
-  /* transform: translate(-103%, 334%); */
-  border-radius: 5px;
-  background-color: #fafafa;
-  box-shadow: 0 0 3px #babbbc;
-  animation: show-profile 0.5s forwards ease-in-out;
-}
-@keyframes show-profile {
-  0% {
-    width: 0;
-  }
-}
-.profile .photo, .profile .content {
-  box-sizing: border-box;
-}
-.profile .photo {
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-  overflow: hidden;
-  border: 2px solid #fafafa;
-  background-color: #fafafa;
-  margin-left: -35px;
-  box-shadow: 0 0 3px #babbbc;
-  animation: rotate-photo 0.5s forwards ease-in-out;
-}
-@keyframes rotate-photo {
-  100% {
-    transform: rotate(-360deg);
-  }
-}
-.profile .photo img {
-  width: 100%;
-}
-.profile .content {
-  padding: 10px;
-  overflow: hidden;
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-}
-.profile .content::before {
-  content: "";
-  position: absolute;
-  width: 158px;
-    height: 81px;
-  background-color: #e9f3e6;
-  left: 0;
-  top: -20px;
-  z-index: -1;
-  transform: rotate(-8deg);
-}
-.profile .content .text {
-    margin-left: 23px;
-}
-.profile .content .text h3, .profile .content .text h6 {
-  font-weight: normal;
-  margin: 3px 0;
-  letter-spacing: 0.5px;
-  white-space: nowrap;
-}
-.profile .content .btn {
-    background-color: #abc;
-    width: 30px;
-    height: 30px;
-    position: absolute;
-    right: 20px;
-    top: 10px;
-  border-radius: 50%;
-  z-index: 1;
-  cursor: pointer;
-  transition-duration: 0.3s;
-  animation: pop-btn 0.3s both ease-in-out 0.5s;
-}
-@keyframes pop-btn {
-  0% {
-    transform: scale(0);
-  }
-  80% {
-    transform: scale(1.2);
-  }
-  100% {
-    transform: scale(1);
-  }
-}
-.profile .content .btn:hover {
-  box-shadow: 0 0 0 5px rgba(170, 187, 204, 0.5);
-}
-.profile .content .btn span {
-  width: 60%;
-  height: 2px;
-  position: absolute;
-  background-color: white;
-  top: 50%;
-  left: 20%;
-  transform: translateY(-50%);
-  animation: to-hamburger 0.3s forwards ease-in-out;
-}
-.profile .content .btn span::before, .profile .content .btn span::after {
-  content: "";
-    width: 79%;
-    height: 2px;
-    position: absolute;
-    background-color: white;
-    transition-duration: 0.3s;
-    transform: rotate(0deg);
-    right: 2px;
-}
-.profile .content .btn span::before {
-  margin-top: -7px;
-}
-.profile .content .btn span::after {
-  margin-top: 7px;
-}
-.profile .content .active span::before {
-  margin-top: -3px;
-}
-.profile .content .active span::after {
-  margin-top: 3px;
-}
-.profile .content .btn.active span {
-  animation: to-arrow 0.3s forwards ease-in-out;
-}
-.profile .content .btn.active span::before, .profile .content .btn.active span::after {
-  width: 60%;
-  right: -1px;
-}
-.profile .content .btn.active span::before {
-  transform: rotate(45deg);
-}
-.profile .content .btn.active span::after {
-  transform: rotate(-45deg);
-}
-@keyframes to-hamburger {
-  from {
-    transform: translateY(-50%) rotate(-180deg);
-  }
-}
-@keyframes to-arrow {
-  from {
-    transform: translateY(-50%) rotate(0deg);
-  }
-  to {
-    transform: translateY(-50%) rotate(180deg);
-  }
-}
-.profile .box {
-  width: 150px;
-  height: 150px;
-  opacity: 0;
-  border-radius: 50%;
-  position: absolute;
-  top: 50%;
-  right: -27%;
-  transform: translate(-50%, -50%);
-  transition-duration: 0.3s;
-}
-.profile .box i {
-  width: 50px;
-  height: 50px;
-  border-radius: 50%;
-  background-color: #ececec;
-  font-size: 26px;
-  text-align: center;
-  line-height: 50px;
-  position: absolute;
-  left: 18px;
-  top: calc(75px - 50px/2);
-  box-shadow: 0 0 0.5rem #babbbc;
-  transition-duration: 0.3s;
-}
-.profile .box i:hover {
-  transition-delay: initial !important;
-  box-shadow: 0 0 0 5px #babbbc;
-}
-.profile .box.open {
-  opacity: 1;
-}
-.profile .box.open i {
-  left: 103px;
-  top: 2em;
-}
-.profile .box.open i:nth-of-type(1) {
-  transform: translateY(2em);
-  transition-delay: 0s;
-}
-.profile .box.open i:nth-of-type(2) {
-  transform: translateY(4em);
-  transition-delay: 0.1s;
-}
-.profile .box.open i:nth-of-type(3) {
-  transform: translateY(6em);
-  transition-delay: 0.2s;
-}
-.profile .box.open i:nth-of-type(4) {
-  transform: translateY(8em);
-  transition-delay: 0.3s;
-}
-.profile .box.open i:nth-of-type(5) {
-  transform: translateY(10em);
-  transition-delay: 0.4s;
-}
