--- Arcano/public/css/footer.css
@@ -1,5 +1,0 @@
-.footer-bottom{
-    width: 100%;
-    padding: 10px 0;
-    text-align: center;
-}
