--- Arcano/public/css/lib/colorPick.css
@@ -1,73 +1,0 @@
-/*!
-*
-* ColorPick jQuery plugin
-* https://github.com/philzet/ColorPick.js
-*
-* Copyright (c) 2017 Phil Zet (a.k.a. Philipp Zakharchenko)
-* Licensed under the MIT License
-*
-*/
-
-@font-face{font-family:'Open Sans';font-style:normal;font-weight:400;src:local('Open Sans'),local('OpenSans'),url(http://fonts.gstatic.com/s/opensans/v13/cJZKeOuBrn4kERxqtaUH3bO3LdcAZYWl9Si6vvxL-qU.woff) format('woff')}@font-face{font-family:'Open Sans';font-style:normal;font-weight:700;src:local('Open Sans Bold'),local('OpenSans-Bold'),url(http://fonts.gstatic.com/s/opensans/v13/k3k702ZOKiLJc3WVjuplzKRDOzjiPcYnFooOUGCOsRk.woff) format('woff')}
-
-#colorPick * {
-	-webkit-transition: all linear .2s;
-	-moz-transition: all linear .2s;
-	-ms-transition: all linear .2s;
-	-o-transition: all linear .2s;
-	transition: all linear .2s;
-}
-
-#colorPick {
-	background: rgba(255, 255, 255, 0.85);
-	-webkit-backdrop-filter: blur(15px);
-	position: absolute;
-	border-radius: 5px;
-	box-shadow: 0px 3px 8px rgba(0, 0, 0, 0.2);
-	padding: 15px;
-	font-family: "Open Sans", sans-serif;
-	width: 140px;
-}
-
-#colorPick span {
-	font-size: 9pt;
-	text-transform: uppercase;
-	font-weight: bold;
-	color: #bbb;
-	margin-bottom: 5px;
-	display: block;
-	clear: both;
-}
-
-.customColorHash {
-	border-radius: 5px;
-	height: 23px;
-	width: 122px;
-	margin: 1px 4px ;
-	padding: 0 4px;
-	border: 1px solid #babbba;	
-	outline: none;
-}
-.customColorHash.error {
-	border-color: #ff424c;
-	color: #ff424c;
-}
-
-.colorPickButton {
-	border-radius: 5px;
-	width: 20px;
-	height: 20px;
-	margin: 0px 3px;
-	cursor: pointer;
-	display: inline-block;
-	border: thin solid #eee;
-}
-
-.colorPickButton:hover {
-	transform: scale(1.1);
-}
-
-.colorPickDummy {
-	background: #fff;
-	border: 1px dashed #bbb;
-}
