--- Arcano/public/js/enigma/index.js
@@ -1,9 +1,0 @@
-var config = {
-    color : '#1ff042',
-    background: '#060e1b',
-    pergunta: 'Eu nunca fui, mas sempre serei. Ninguém nunca me viu, e nunca verão. Ainda assim, sou a esperança de todos. Quem sou eu?',
-    resposta : 'amanha',
-    dica:  'Não é Hoje'
-}
-new Controller($('.404-input'),config);
-
