--- Arcano/src/Controllers/Cadastro.php
@@ -1,13 +1,0 @@
-<?php
-
-
-namespace Arcanos\Enigmas\Controllers;
-
-
-class Cadastro extends Banco implements RequestHandlerInterface
-{
-    public function handle()
-    {
-        $this->view('Cadastro.php',[]);
-    }
-}
