--- Arcano/src/Controllers/Home.php
@@ -1,13 +1,0 @@
-<?php
-
-
-namespace Arcanos\Enigmas\Controllers;
-
-
-class Home extends Banco implements RequestHandlerInterface
-{
-    public function handle()
-    {
-        $this->view('Home.php',[]);
-    }
-}
