--- Arcano/src/Controllers/Login.php
@@ -1,14 +1,0 @@
-<?php
-
-
-namespace Arcanos\Enigmas\Controllers;
-
-
-class Login extends Banco implements RequestHandlerInterface
-{
-
-    public function handle()
-    {
-        $this->view('login.php',[]);
-    }
-}
