--- Arcano/src/Controllers/RequestHandlerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-
-namespace Arcanos\Enigmas\Controllers;
-
-
-interface RequestHandlerInterface
-{
-    public function handle();
-}
