--- Arcano/view/footer.php
@@ -1,6 +1,0 @@
-        </div>
-        <div class="footer-bottom fixed-bottom bg-light">
-            Copyright © 2017, All Rights Reserved
-        </div>
-    </body>
-</html>
