--- Arcano/view/pages/Enigma.php
@@ -1,17 +1,0 @@
-<link rel="stylesheet" href="css/enigma.css">
-
-<div class="container-enigma" id="myElement">
-    <form class="four-oh-four-form">
-        <input type="text" class="404-input">
-    </form>
-    <div class="terminal">
-        <p class="prompt question"></p>
-        <p class="prompt output new-output"></p>
-    </div>
-</div>
-<div class="picker picker-background"></div>
-<div class="picker picker-color"></div>
-
-<script  src="js/enigma/Engine.js"></script>
-<script  src="js/enigma/Controller.js"></script>
-<script  src="js/enigma/index.js"></script>
