--- Arcano/view/pages/Home.php
@@ -1,28 +1,0 @@
-<link rel="stylesheet" type="text/css" href="css/home.css">
-<link rel="stylesheet" type="text/css" href="css/testeBG.css">
-<style>
-.container{
-    margin: 0px;
-    padding: 0px;
-}
-body {
-    background: #060e1b;
-    overflow: hidden;
-}
-
-#canvas{
-    position: absolute;
-    top: 0;
-    left: 0;
-}
-</style>
-    <div>
-        <canvas id="canvas"></canvas>
-        <script  src="js/scriptBG.js"></script>
-        <div class="logoCirculo">
-            <img src="img/Circulologo.png">
-        </div>
-        <div class="logoPrincipal">
-            <img src="img/Logo.png">
-        </div>
-    </div>
