--- Arcano/view/pages/login.php
@@ -1,15 +1,0 @@
-
-    <form action="" method="GET">
-      <div class="form-row">
-        <div class="form-group col-md-6">
-            <input type="text" class="form-control" name="usuario" placeholder="Usuario">
-        </div>
-        <div class="form-group col-md-6">
-            <input type="password" class="form-control" name="senha" placeholder="Senha">
-        </div>
-        <div class="w-100"></div>
-        <div class="form-group col-md-6">
-            <button type="submit" class="btn btn-primary">Entrar</button>
-        </div>
-      </div>
-    </form>
