+++ src/Controllers/PersistenciaUsuario.php
@@ -14,13 +14,45 @@ class PersistenciaUsuario extends Banco implements RequestHandlerInterface
             $senha = $_POST['senha'];
             $confsenha = $_POST['confsenha'];
 
-            if ($senha < 4) {
-               ;
+            $sla = array($this->banco->select("SELECT * FROM usuarios",[2],true));
+
+            $validar = array($sla);
+            //$arrValidar = mysqli_fetch_all($validar, MYSQLI_ASSOC);
+
+            if(filter_var($email, FILTER_VALIDATE_EMAIL))
+            {
+                
+            }
+            else
+            {
+                header("Location: index.php?pagina=cadastro&msg=1");
             }
 
+            foreach ($validar as $key => $value) {
+                if ($value['nome_usuario'] == $validar ) {
+                    
+                    header("Location: index.php?pagina=cadastro&msg=2");
+                    
+                }
+                elseif ($value['email'] == $email) {
+                    header("Location: index.php?pagina=cadastro&msg=3");
+                }
+            }
+
+            if ($senha != $confsenha) {
+
+
+                header("Location: index.php?pagina=cadastro&msg=4");
+               
+            }
+            elseif (mb_strlen($senha) <6) {
+                
+                header("Location: index.php?pagina=cadastro&msg=5");
+            }
+            //criar função para validar senha, na barra de favoritos tem sites para ajudar
             //$retorno = $this->usuarios->cadastrarUsuario($nome_usuario,2,$email,$senha)
         
-            echo $usuario, $email, $senha, $confsenha;
+            //echo $usuario, $email, $senha, $confsenha;
             
         }
     }
